requirements.txt
aiogram
pandas
numpy
scikit-learn
statsmodels
matplotlib
seaborn
python-dotenv
optuna
requests
tensorflow
aiofiles